#80 [REFACTOR] CustomAnalysisEntity 관련 객체들 domain으로 이전
 - 판단하는 비즈니스 로직을 포함하고 있어 이전하였습니다.
