+++ src/main/java/com/wakeUpTogetUp/togetUp/api/mission/MissionService.java
@@ -6,8 +6,7 @@ import com.wakeUpTogetUp.togetUp.api.alarm.AlarmRepository;
 import com.wakeUpTogetUp.togetUp.api.alarm.model.Alarm;
 import com.wakeUpTogetUp.togetUp.api.mission.dto.request.MissionCompleteReq;
 import com.wakeUpTogetUp.togetUp.api.mission.dto.response.MissionCompleteRes;
-import com.wakeUpTogetUp.togetUp.api.mission.model.CustomAnalysisEntity;
-import com.wakeUpTogetUp.togetUp.api.mission.model.Emotion;
+import com.wakeUpTogetUp.togetUp.api.mission.domain.CustomAnalysisEntity;
 import com.wakeUpTogetUp.togetUp.api.mission.model.MissionLog;
 import com.wakeUpTogetUp.togetUp.api.mission.domain.MissionPerformResult;
 import com.wakeUpTogetUp.togetUp.infra.azure.vision.AzureVision32Service;
